Conflicts:
	cpp/game/simulator.h
	cpp/schedulers/bulk_scheduler.cpp
	cpp/schedulers/shortest_path_switch_scheduler.cpp
	cpp/schedulers/wojtek_throttle_scheduler.cpp

+++ cpp/game/double_path_optimizer.cpp
@@ -11,7 +11,7 @@ DoublePathOptimizer::DoublePathOptimizer(const Race& race, CarTracker& car_track
   optimizer_started_ = false;
 }
 
-std::map<Switch, int> DoublePathOptimizer::Score(const Position& position) {
+std::map<Switch, double> DoublePathOptimizer::Score(const Position& position) {
   if (is_perfect_ready_.load() == true) {
     return perfect_.Score(position);
   } else {
