+++ cpp/game/greedy_path_optimizer.h
@@ -12,7 +12,7 @@ class GreedyPathOptimizer : public PathOptimizerInterface {
   // Returns all possible decisions with associated score
   // score means the loss of time (lap-wise) compared to optimal choice
   // so atleast one lane will always have value of 0
-  std::map<Switch, int> Score(const Position& position) override;
+  std::map<Switch, double> Score(const Position& position) override;
 
  private:
   // The distance between 'start' and 'end' assuming that 'end' position is on
